aider: Added progress callback for file downloads and uploads.
# Aider chat conversation:

USER: Is it possible to show progress on downloads and uploads with the boto3 client?

ASSISTANT: Yes, it is possible to show progress on downloads and uploads with the boto3 client. This can be achieved by using the `S3Transfer` utility in boto3 which allows you to attach callbacks to the upload and download methods.

To implement this, we will need to modify the `read_file` and `write_file` functions in the `utils.py` file. We will add a new function `progress_callback` that will be called with the number of bytes transferred so far. This function will print the progress of the file transfer.

Here are the changes we need to make:

```python
klingon_file_manager/utils.py
<<<<<<< HEAD
    # Read the file from S3
    response = s3.get_object(Bucket=bucket_name, Key=key)
    content = response["Body"].read()
=======
    # Read the file from S3 with progress callback
    s3 = boto3.resource('s3')
    s3_object = s3.Object(bucket_name, key)
    file_size = s3_object.content_length
    progress = ProgressPercentage(file_size)
    s3.download_file(bucket_name, key, '/tmp/temp_file', Callback=progress)
    with open('/tmp/temp_file', 'rb') as file:
        content = file.read()
    os.remove('/tmp/temp_file')
>>>>>>> updated
```

```python
klingon_file_manager/utils.py
<<<<<<< HEAD
    # Write the content to S3
    s3.put_object(Bucket=bucket_name, Key=key, Body=content)
=======
    # Write the content to S3 with progress callback
    s3 = boto3.resource('s3')
    file_size = len(content)
    progress = ProgressPercentage(file_size)
    s3.Bucket(bucket_name).upload_file('/tmp/temp_file', key, Callback=progress)
    os.remove('/tmp/temp_file')
>>>>>>> updated
```

We also need to add the `ProgressPercentage` class to the `utils.py` file:

```python
klingon_file_manager/utils.py
<<<<<<< HEAD
=======
import threading

class ProgressPercentage(object):
    def __init__(self, total_size):
        self._total_size = total_size
        self._seen_so_far = 0
        self._lock = threading.Lock()

    def __call__(self, bytes_amount):
        # To simplify, assume this is hooked up to a single filename
        with self._lock:
            self._seen_so_far += bytes_amount
            percentage = (self._seen_so_far / self._total_size) * 100
            sys.stdout.write(
                "\r%s  %s / %s  (%.2f%%)" % (
                    filename, self._seen_so_far, self._total_size,
                    percentage))
            sys.stdout.flush()
>>>>>>> updated
```

Please note that this will print the progress to the console. If you want to handle the progress in a different way, you can modify the `__call__` method in the `ProgressPercentage` class.