aider: Create temporary file and write content to it before uploading to S3.
# Aider chat conversation:

ASSISTANT: The error message indicates that the tests `test_post_s3_txt_file` and `test_post_s3_bin_file` are failing because the file `/tmp/temp_file` does not exist. This file is being used as a temporary file in the `write_file` function in `klingon_file_manager/write.py` when writing to an S3 bucket.

The issue seems to be that the content is being written to the S3 bucket via a temporary file, but the temporary file is not being created before the write operation is attempted.

To fix this, we need to create the temporary file and write the content to it before attempting to upload it to the S3 bucket.

The files to modify are:

1. `klingon_file_manager/write.py`

Here are the changes:

```python
klingon_file_manager/write.py
<<<<<<< HEAD
                # Write the content to S3 with progress callback
                s3 = boto3.resource('s3')
                file_size = len(content)
                progress = ProgressPercentage(file_size)
                s3.Bucket(bucket_name).upload_file('/tmp/temp_file', key, Callback=progress)
                os.remove('/tmp/temp_file')
=======
                # Write the content to a temporary file
                with open('/tmp/temp_file', 'wb') as temp_file:
                    temp_file.write(content)
                # Write the content to S3 with progress callback
                s3 = boto3.resource('s3')
                file_size = len(content)
                progress = ProgressPercentage(file_size)
                s3.Bucket(bucket_name).upload_file('/tmp/temp_file', key, Callback=progress)
                os.remove('/tmp/temp_file')
>>>>>>> updated
```

This change will create the temporary file and write the content to it before attempting to upload it to the S3 bucket.