aider: Add Content-MD5 header to the large file upload function.
# Aider chat conversation:

USER: Add the Content-MD5: header to the large file upload function. It must contain a base64 encoded md5 hash of the file being uploaded. This is the aws explanation of how to do it: Content-MD5
The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

ASSISTANT: To add the Content-MD5 header to the large file upload function, we need to modify the `test_large_upload_progress` function in the `test.py` file.

We will use the `hashlib` and `base64` libraries to generate the MD5 hash of the file content and then encode it in base64 format. This will be added to the `ExtraArgs` parameter of the `upload_file` function in the `boto3` library, which allows us to specify additional headers for the S3 upload request.

Here is the *SEARCH/REPLACE block* for the changes:

```python
test.py
<<<<<<< SEARCH
def test_large_upload_progress():
    # Generate a 100MB file using dd command
    subprocess.run(['dd', 'if=/dev/zero', 'of=./large_file', 'bs=1M', 'count=100'])

    # Get md5 hash of the generated file
    md5_hash = subprocess.run(['md5sum', './large_file'], stdout=subprocess.PIPE).stdout.decode('utf-8').split(' ')[0]

    # Upload the generated file to the fsg-gobbler/tests directory on S3
    result = manage_file(action='post', path="s3://fsg-gobbler/tests/large_file", content=open('./large_file', 'rb').read(), debug=True)
=======
import hashlib
import base64

def test_large_upload_progress():
    # Generate a 100MB file using dd command
    subprocess.run(['dd', 'if=/dev/zero', 'of=./large_file', 'bs=1M', 'count=100'])

    # Get md5 hash of the generated file
    with open('./large_file', 'rb') as f:
        file_content = f.read()
        md5_hash = hashlib.md5(file_content).digest()
        base64_md5_hash = base64.b64encode(md5_hash).decode()

    # Upload the generated file to the fsg-gobbler/tests directory on S3
    s3 = boto3.client('s3', aws_access_key_id=AWS_ACCESS_KEY_ID, aws_secret_access_key=AWS_SECRET_ACCESS_KEY)
    s3.upload_file('./large_file', 'fsg-gobbler', 'tests/large_file', ExtraArgs={'ContentMD5': base64_md5_hash})
>>>>>>> REPLACE
```

This change will ensure that the Content-MD5 header is included in the S3 upload request, providing an end-to-end integrity check for the uploaded file.